fix: don't install windows .dotfiles auto

+++ install-dotfiles.sh
@@ -15,11 +15,20 @@ for DOTFILE in $(find -H $DOTS -maxdepth 2 -type f)
 do
   filename=$(basename "${DOTFILE}")
   next "copy $filename"
+
   if [[ -e "$HOME/$filename" ]]; then
     mv "$HOME/$filename" "$HOME/$filename.old"
+  else 
+
+  fi
+
+  if [[ $DOTFILE =~ "windows" || -L "$HOME/$filename" ]]; then
+    continue
   fi
   
   ln -s $DOTFILE "$HOME/$filename"
 
   check "$filename configured"
 done
+
+check "Configuration done. You might need to restart your terminal to finish the installation"
